+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Add confirmation page when publishing decision data from the editing page
+
 ## [release-019] - 2022-05-04
 
 ### Changed